change: to using bigints for ids in db

+++ db/migrate/20190509125739_change_likes_to_polymorphic.rb
@@ -3,7 +3,7 @@ class ChangeLikesToPolymorphic < ActiveRecord::Migration[5.2]
     remove_index  'likes', name: 'index_likes_on_post_id'
     remove_column :likes, :post_id
 
-    add_column    :likes, :likeable_id,   :integer
+    add_column    :likes, :likeable_id,   :bigint
     add_column    :likes, :likeable_type, :string
     add_index     :likes, [:likeable_type, :likeable_id]
   end
